+++ api/src/test/scala/hmda/api/http/institutions/submissions/SubmissionBasePathsSpec.scala
@@ -1,4 +1,4 @@
-package hmda.api.http.institutions
+package hmda.api.http.institutions.submissions
 
 import akka.actor.ActorRef
 import akka.http.scaladsl.marshallers.sprayjson.SprayJsonSupport._
@@ -16,14 +16,9 @@ import hmda.validation.engine._
 
 import scala.concurrent.Future
 
-class SubmissionPathsSpec extends InstitutionHttpApiSpec {
+class SubmissionBasePathsSpec extends InstitutionHttpApiSpec {
 
   val supervisor = system.actorSelection("/user/supervisor")
-
-  override def beforeAll(): Unit = {
-    super.beforeAll()
-    loadValidationErrors()
-  }
 
   "Submission Paths" must {
     "find the latest submission for an institution" in {
@@ -72,67 +67,5 @@ class SubmissionPathsSpec extends InstitutionHttpApiSpec {
     postWithCfpbHeaders("/institutions/0/filings/2017/submissions/latest") ~> institutionsRoutes ~> check {
       rejection mustBe a[MethodRejection]
     }
-  }
-
-  "return summary of validation errors" in {
-    val expectedSummary = SummaryEditResults(
-      EditResults(
-        List(
-          EditResult("S020", List(LarEditResult(LarId("loan1")))),
-          EditResult("S010", List(LarEditResult(LarId("loan1"))))
-        )
-      ),
-      EditResults(
-        List(
-          EditResult("V285", List(LarEditResult(LarId("loan2")), LarEditResult(LarId("loan3")))),
-          EditResult("V280", List(LarEditResult(LarId("loan1"))))
-        )
-      ),
-      EditResults.empty,
-      EditResults.empty
-    )
-
-    getWithCfpbHeaders(s"/institutions/0/filings/2017/submissions/1/edits") ~> institutionsRoutes ~> check {
-      status mustBe StatusCodes.OK
-      responseAs[SummaryEditResults] mustBe expectedSummary
-    }
-  }
-
-  "return a list of validation errors for a single type" in {
-    val expectedEdits =
-      EditResults(
-        List(
-          EditResult("V285", List(LarEditResult(LarId("loan2")), LarEditResult(LarId("loan3")))),
-          EditResult("V280", List(LarEditResult(LarId("loan1"))))
-        )
-      )
-
-    getWithCfpbHeaders(s"/institutions/0/filings/2017/submissions/1/edits/validity") ~> institutionsRoutes ~> check {
-      status mustBe StatusCodes.OK
-      responseAs[EditResults] mustBe expectedEdits
-    }
-  }
-
-  private def loadValidationErrors(): Unit = {
-    val supervisor = system.actorSelection("/user/supervisor")
-    val id = "0"
-    val period = "2017"
-    val seqNr = 1
-    val submissionId = SubmissionId(id, period, seqNr)
-    val fHmdaValidator = (supervisor ? FindProcessingActor(HmdaFileValidator.name, submissionId)).mapTo[ActorRef]
-
-    val s1 = ValidationError("loan1", "S010", Syntactical)
-    val s2 = ValidationError("loan1", "S020", Syntactical)
-    val v1 = ValidationError("loan1", "V280", Validity)
-    val v2 = ValidationError("loan2", "V285", Validity)
-    val v3 = ValidationError("loan3", "V285", Validity)
-    val validationErrors = LarValidationErrors(Seq(s1, s2, v1, v2, v3))
-
-    val fValidate: Future[Unit] = for {
-      h <- fHmdaValidator
-    } yield {
-      h ! validationErrors
-    }
-
   }
 }
